Move variable conversion code to module
