+++ src/met-office/CMakeLists.txt
@@ -1,4 +1,5 @@
 install (PROGRAMS
   odbapi2nc.py
   odbapi2json.py
+  var_convert.py
   DESTINATION bin)
